requirements.txt
streamlit
rembg[cpu]
opencv-python-headless
numpy
pillow